+++ examples/plugins/log-plugin/main.go
@@ -1,4 +1,4 @@
-// Copyright (c) 2020, Sylabs Inc. All rights reserved.
+// Copyright (c) 2020-2021, Sylabs Inc. All rights reserved.
 // This software is licensed under a 3-clause BSD license. Please consult the
 // LICENSE.md file distributed with the sources of this project regarding your
 // rights to use or distribute this software.
@@ -9,12 +9,12 @@ import (
 	"fmt"
 	"log/syslog"
 	"os"
-	"path/filepath"
 
+	"github.com/spf13/cobra"
+	"github.com/sylabs/singularity/pkg/cmdline"
 	pluginapi "github.com/sylabs/singularity/pkg/plugin"
-	singularitycallback "github.com/sylabs/singularity/pkg/plugin/callback/runtime/engine/singularity"
-	"github.com/sylabs/singularity/pkg/runtime/engine/config"
-	singularityConfig "github.com/sylabs/singularity/pkg/runtime/engine/singularity/config"
+	clicallback "github.com/sylabs/singularity/pkg/plugin/callback/cli"
+	"github.com/sylabs/singularity/pkg/sylog"
 )
 
 // Plugin is the only variable which a plugin MUST export.
@@ -23,31 +23,44 @@ var Plugin = pluginapi.Plugin{
 	Manifest: pluginapi.Manifest{
 		Name:        "github.com/sylabs/singularity/log-plugin",
 		Author:      "Sylabs Team",
-		Version:     "0.1.0",
-		Description: "Log executed commands to syslog",
+		Version:     "0.2.0",
+		Description: "Log executed CLI commands to syslog",
 	},
 	Callbacks: []pluginapi.Callback{
-		(singularitycallback.PostStartProcess)(logCommand),
+		(clicallback.Command)(logCommand),
 	},
 }
 
-func logCommand(common *config.Common, pid int) error {
-	cfg := common.EngineConfig.(*singularityConfig.EngineConfig)
+func logCommand(manager *cmdline.CommandManager) {
+	rootCmd := manager.GetRootCmd()
 
-	command := "unknown"
-	if cfg.OciConfig != nil && cfg.OciConfig.Process != nil {
-		if len(cfg.OciConfig.Process.Args) > 0 {
-			command = filepath.Base(cfg.OciConfig.Process.Args[0])
+	// Keep track of an existing PreRunE so we can call it
+	f := rootCmd.PersistentPreRunE
+
+	// The log action is added as a PreRunE on the main `singularity` root command
+	// so we can log anything a user does with `singularity`.
+	rootCmd.PersistentPreRunE = func(c *cobra.Command, args []string) error {
+		uid := os.Getuid()
+		gid := os.Getgid()
+		command := c.Name()
+		msg := fmt.Sprintf("UID=%d GID=%d COMMAND=%s ARGS=%v", uid, gid, command, args)
+
+		// This logger never errors, only warns, if it fails to write to syslog
+		w, err := syslog.New(syslog.LOG_INFO, "singularity")
+		if err != nil {
+			sylog.Warningf("Could not create syslog: %v", err)
+		} else {
+			defer w.Close()
+			if err := w.Info(msg); err != nil {
+				sylog.Warningf("Could not write to syslog: %v", err)
+			}
 		}
-	}
 
-	image := cfg.GetImage()
-	w, err := syslog.New(syslog.LOG_INFO, "singularity")
-	if err != nil {
-		return err
-	}
-	defer w.Close()
+		// Call any existing PreRunE
+		if f != nil {
+			return f(c, args)
+		}
 
-	msg := fmt.Sprintf("UID=%d IMAGE=%s COMMAND=%s", os.Getuid(), image, command)
-	return w.Info(msg)
+		return nil
+	}
 }